Add zooming and panning to fractal

+++ src/Hsl.hpp
@@ -5,11 +5,11 @@
 
 class Hsl {
   public:
-    uint16_t h;
+    double h;
     double s;
     double l;
 
-    Hsl(uint16_t hue, double sat, double light) : h(hue), s(sat), l(light) {};
+    Hsl(double hue, double sat, double light) : h(hue), s(sat), l(light) {};
     Uint32 toRgb();
 
   private:
